Added the route and controller to set the new password of an user
